requirements.txt

reflex==0.8.17
google-generativeai
gql
requests
requests-toolbelt